update fasta parsing to be configurable, and add tests for new settings
also removes the distinction between a DnaFastaParser and a "regular"
SimpleFastaParser -- everything is "FastaParser" now, with a phantom
generic FromStr type that will be parsed (String has an infalliable FromStr impl)
So SimpleFastaParser is now FastaParser<String>, and DnaFastaParser is now
FastaParser<DnaSequence>. ProteinFastaParser didn't exist before, but it's now
FastaParser<ProteinSequence>. A hypothetical integer fasta format could be
FastaParser<u64>, if you wanted that for some weird reason.

Fasta parsing is now state-machine-driven, which should make a non-allocating
API easier to add in the future, and also simplified the parsing code since it was
getting hairy handling the new special cases in the old loop.