add tests to fully match the spec (#12)
Our spec is as follows:
```
The FASTA format is poorly standardized; see the URLs on this page for pointers to several partial and conflicting definitions. We adopt a consensus view and try to be liberal in what we accept.
One record consists one or more consecutive header lines, followed by one or (typically, many more) DNA nucleotides.
There can be any number of these records concatenated into a single string; we will screen them all.
Header lines MUST have > or the semi-obsolete ; (0x3b) in the first column to be recognized as such.
Header lines MAY use UTF-8, although > or ; MUST be ASCII characters (0x3e or 0x3b).
Header lines MAY be nonunique. In other words, the same header line MAY appear in more than one place in the input.
Header lines are ignored for screening purposes but are used when identifying hazard hits for customer convenience.
Any text present in the input before the first header line treated as if it is sequence data; it is not ignored. In other words, in this case, the very first record MAY have zero header lines associated with it. (Some FASTA files appear to treat this as a comment, but we cannot, because we have no guarantee that all providers will do so. Any provider which treated a headerless sequence as a synthesis request could therefore allow a trivial screening bypass if we ignored this text; customers with such files SHOULD be encouraged to fix them via the provider checking their input and complaining before attempting to screen.)
Multiple header lines MAY appear with no intervening sequence; if so, they are treated as if all of them describe the following sequence.
The sequence information itself MAY be on separate lines of any length, or on one long single line, with no line-length limit.
Sequence information MUST use only ASCII characters. Characters outside of allowable DNA nucleotides, plus ASCII whitespace, will cause the request to be rejected.
Line termination MAY be ASCII newline ('\n', 0x0a), ASCII carriage return ('\r', 0x0d), or both ('\r\n', x0a0d).
The input string is NOT REQUIRED to end a line termination character, although it is likely that it will.
Alphabetic case in sequences is ignored.
```

Some of the cases were not covered by our tests. Add them. Luckily, no code changes were required.